+++ interfaces/kits/gfx_utils/graphics/transform/graphic_transform_affine.h
@@ -20,7 +20,7 @@
 #include "gfx_utils/graphics/common/graphic_common_basics.h"
 
 namespace OHOS {
-    const double affineEpsilon = 1e-14;
+    const float affineEpsilon = 1e-14;
 
     /**
      * @brief 图源映射变换
@@ -28,7 +28,7 @@ namespace OHOS {
      * @version 1.0
      */
     struct TransAffine {
-        double scaleX_, shearY, shearX, scaleY_, translateX, translateY;
+        float scaleX_, shearY, shearX, scaleY_, translateX, translateY;
         /**
          * @brief 初始化为单位矩阵
          * @since 1.0
@@ -44,7 +44,7 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        TransAffine(double v0, double v1, double v2, double v3, double v4, double v5)
+        TransAffine(float v0, float v1, float v2, float v3, float v4, float v5)
             : scaleX_(v0), shearY(v1),
               shearX(v2), scaleY_(v3),
               translateX(v4), translateY(v5)
@@ -54,7 +54,7 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        TransAffine(double x1, double y1, double x2, double y2, const double* parl)
+        TransAffine(float x1, float y1, float x2, float y2, const float* parl)
         {
             RectToParl(x1, y1, x2, y2, parl);
         }
@@ -66,7 +66,7 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        const TransAffine& ParlToParl(const double* src, const double* dst);
+        const TransAffine& ParlToParl(const float* src, const float* dst);
         /**
          * @brief 将原矩形转换为目标平行四边形
          * @param x1 矩形左上角x轴坐标
@@ -77,9 +77,9 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        const TransAffine& RectToParl(double x1, double y1,
-                                      double x2, double y2,
-                                      const double* parl);
+        const TransAffine& RectToParl(float x1, float y1,
+                                      float x2, float y2,
+                                      const float* parl);
 
         /**
          * @brief 重置矩阵
@@ -95,21 +95,21 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        const TransAffine& Translate(double x, double y);
+        const TransAffine& Translate(float x, float y);
         /**
          * @brief 图元旋转
          * @param angle 旋转角度
          * @since 1.0
          * @version 1.0
          */
-        const TransAffine& Rotate(double angle);
+        const TransAffine& Rotate(float angle);
         /**
          * @brief 图元放大
          * @param scale 整体放大系数
          * @since 1.0
          * @version 1.0
          */
-        const TransAffine& Scale(double scale);
+        const TransAffine& Scale(float scale);
         /**
          * @brief 图元放大
          * @param x x轴放大系数
@@ -117,7 +117,7 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        const TransAffine& Scale(double x, double y);
+        const TransAffine& Scale(float x, float y);
 
         /**
          * @brief 转换矩阵相乘
@@ -146,7 +146,7 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        void Transform(double* x, double* y) const;
+        void Transform(float* x, float* y) const;
 
         /**
          * @brief 逆变换
@@ -155,14 +155,14 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        void InverseTransform(double* x, double* y) const;
+        void InverseTransform(float* x, float* y) const;
 
         /**
          * @brief 计算行列式的倒数
          * @since 1.0
          * @version 1.0
          */
-        double DeterminantReciprocal() const
+        float DeterminantReciprocal() const
         {
             return 1.0 / (scaleX_ * scaleY_ - shearY * shearX);
         }
@@ -172,14 +172,14 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        bool IsValid(double epsilon = affineEpsilon) const;
+        bool IsValid(float epsilon = affineEpsilon) const;
 
         /**
          * @brief 检查它是否是单位矩阵
          * @since 1.0
          * @version 1.0
          */
-        bool IsIdentity(double epsilon = affineEpsilon) const;
+        bool IsIdentity(float epsilon = affineEpsilon) const;
 
         /**
          * @brief 缩放
@@ -188,37 +188,37 @@ namespace OHOS {
          * @since 1.0
          * @version 1.0
          */
-        void ScalingAbs(double* x, double* y) const;
+        void ScalingAbs(float* x, float* y) const;
     };
 
-    inline void TransAffine::Transform(double* x, double* y) const
+    inline void TransAffine::Transform(float* x, float* y) const
     {
-        double tmp = *x;
+        float tmp = *x;
         *x = tmp * scaleX_ + *y * shearX + translateX;
         *y = tmp * shearY + *y * scaleY_ + translateY;
     }
 
-    inline void TransAffine::InverseTransform(double* x, double* y) const
+    inline void TransAffine::InverseTransform(float* x, float* y) const
     {
-        double reciprocal = DeterminantReciprocal();
-        double a = (*x - translateX) * reciprocal;
-        double b = (*y - translateY) * reciprocal;
+        float reciprocal = DeterminantReciprocal();
+        float a = (*x - translateX) * reciprocal;
+        float b = (*y - translateY) * reciprocal;
         *x = a * scaleY_ - b * shearX;
         *y = b * scaleX_ - a * shearY;
     }
 
-    inline const TransAffine& TransAffine::Translate(double deltaX, double deltaY)
+    inline const TransAffine& TransAffine::Translate(float deltaX, float deltaY)
     {
         translateX += deltaX;
         translateY += deltaY;
         return *this;
     }
 
-    inline const TransAffine& TransAffine::Rotate(double angle)
+    inline const TransAffine& TransAffine::Rotate(float angle)
     {
-        double scaleXTemp = scaleX_ * std::cos(angle) - shearY * std::sin(angle);
-        double shearXTemp = shearX * std::cos(angle) - scaleY_ * std::sin(angle);
-        double translateXTemp = translateX * std::cos(angle) - translateY * std::sin(angle);
+        float scaleXTemp = scaleX_ * std::cos(angle) - shearY * std::sin(angle);
+        float shearXTemp = shearX * std::cos(angle) - scaleY_ * std::sin(angle);
+        float translateXTemp = translateX * std::cos(angle) - translateY * std::sin(angle);
         shearY = scaleX_ * std::sin(angle) + shearY * std::cos(angle);
         scaleY_ = shearX * std::sin(angle) + scaleY_ * std::cos(angle);
         translateY = translateX * std::sin(angle) + translateY * std::cos(angle);
@@ -228,10 +228,10 @@ namespace OHOS {
         return *this;
     }
 
-    inline const TransAffine& TransAffine::Scale(double scaleX, double scaleY)
+    inline const TransAffine& TransAffine::Scale(float scaleX, float scaleY)
     {
-        double mm0 = scaleX;
-        double mm3 = scaleY;
+        float mm0 = scaleX;
+        float mm3 = scaleY;
         scaleX_ *= mm0;
         shearX *= mm0;
         translateX *= mm0;
@@ -241,9 +241,9 @@ namespace OHOS {
         return *this;
     }
 
-    inline const TransAffine& TransAffine::Scale(double scale)
+    inline const TransAffine& TransAffine::Scale(float scale)
     {
-        double m = scale;
+        float m = scale;
         scaleX_ *= m;
         shearX *= m;
         translateX *= m;
@@ -253,7 +253,7 @@ namespace OHOS {
         return *this;
     }
 
-    inline void TransAffine::ScalingAbs(double* x, double* y) const
+    inline void TransAffine::ScalingAbs(float* x, float* y) const
     {
         *x = std::sqrt(scaleX_ * scaleX_ + shearX * shearX);
         *y = std::sqrt(shearY * shearY + scaleY_ * scaleY_);
@@ -266,7 +266,7 @@ namespace OHOS {
      */
     class TransAffineRotation : public TransAffine {
     public:
-        TransAffineRotation(double angle)
+        TransAffineRotation(float angle)
             : TransAffine(std::cos(angle), std::sin(angle), -std::sin(angle), std::cos(angle), 0.0, 0.0)
         {}
     };
@@ -278,10 +278,10 @@ namespace OHOS {
      */
     class TransAffineScaling : public TransAffine {
     public:
-        TransAffineScaling(double x, double y) : TransAffine(x, 0.0, 0.0, y, 0.0, 0.0)
+        TransAffineScaling(float x, float y) : TransAffine(x, 0.0, 0.0, y, 0.0, 0.0)
         {}
 
-        TransAffineScaling(double scale) : TransAffine(scale, 0.0, 0.0, scale, 0.0, 0.0)
+        TransAffineScaling(float scale) : TransAffine(scale, 0.0, 0.0, scale, 0.0, 0.0)
         {}
     };
     /**
@@ -291,7 +291,7 @@ namespace OHOS {
      */
     class TransAffineTranslation : public TransAffine {
     public:
-        TransAffineTranslation(double x, double y) : TransAffine(1.0, 0.0, 0.0, 1.0, x, y)
+        TransAffineTranslation(float x, float y) : TransAffine(1.0, 0.0, 0.0, 1.0, x, y)
         {}
     };
 } // namespace OHOS
